requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
pydantic==2.11.4
gunicorn==21.2.0

# 핵심 모델 관련
transformers==4.51.3
torch==2.6.0
accelerate==1.6.0
safetensors==0.5.3
sentencepiece==0.1.99
peft==0.8.2
einops==0.7.0
bitsandbytes==0.42.0

# RAG 및 벡터 검색 관련
faiss-cpu==1.11.0
sentence-transformers==4.1.0
PyMuPDF==1.23.22
langchain==0.3.25

# 기타
protobuf==5.28.1
numpy==1.26.4
python-dotenv==1.0.1
tokenizers==0.21.1
huggingface_hub==0.30.1
pyngrok==6.0.0